content: update TOB membership
It seems as though during the rewrite, 3 of the TOB members were
omitted -- correct this mistake and clean up the TOB member descriptions
(add an image for all members I could find an image of, and include the
term dates and roles for each member).

Ideally each member would have a full biography, but that's up to them
to write.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>